+++ pepselect-child/inc/setup.php
@@ -26,6 +26,7 @@ require_once get_stylesheet_directory() . '/inc/account.php';
 require_once get_stylesheet_directory() . '/inc/referral-vanity.php';
 require_once get_stylesheet_directory() . '/inc/checkout.php';
 require_once get_stylesheet_directory() . '/inc/checkout-fields.php';
+require_once get_stylesheet_directory() . '/inc/cart-empty.php';
 require_once get_stylesheet_directory() . '/inc/legacy-shortcodes.php';
 require_once get_stylesheet_directory() . '/inc/emails.php';